refactor: Move AWS key constants into standalone file
Avoids the CLI having to import an unused jsonschema package.
