+++ tests/test_dataset_versions_endpoint.py
@@ -12,7 +12,7 @@ from pytest import mark
 from pytest_subtests import SubTests
 
 from geostore.api_keys import MESSAGE_KEY
-from geostore.api_responses import BODY_KEY, HTTP_METHOD_KEY, STATUS_CODE_KEY
+from geostore.aws_keys import BODY_KEY, HTTP_METHOD_KEY, STATUS_CODE_KEY
 from geostore.dataset_versions import entrypoint
 from geostore.dataset_versions.create import create_dataset_version
 from geostore.step_function_keys import (
